gcb: Moved tests into separate riscv64 package
